Prevent link activation by enter key
Without this commit, if a link is focused when the user types a number
followed by enter, the slide first changes based on the jump plugin.
Immediately afterwards, the link is activated by the enter key to jump
elsewhere.
With this commit, the enter key is only used to jump once.